+++ app/src/main/java/com/konkuk/moru/presentation/navigation/MainNavGraph.kt
@@ -32,7 +32,7 @@ import com.konkuk.moru.presentation.myactivity.screen.ActScrabScreen
 import com.konkuk.moru.presentation.myactivity.screen.ActSettingScreen
 import com.konkuk.moru.presentation.myroutines.screen.MyRoutineDetailScreen
 import com.konkuk.moru.presentation.myroutines.screen.MyRoutinesScreen
-import com.konkuk.moru.presentation.myroutines.screen.MyRoutinesViewModel
+import com.konkuk.moru.presentation.myroutines.viewmodel.MyRoutinesViewModel
 import com.konkuk.moru.presentation.routinecreate.screen.RoutineCreateScreen
 import com.konkuk.moru.presentation.routinefeed.screen.NotificationScreen
 import com.konkuk.moru.presentation.routinefeed.screen.follow.FollowScreen
@@ -217,8 +217,6 @@ fun MainNavGraph(
                 title == "지금 가장 핫한 루틴은?" -> feedRoutines.filter { it.likes > 70 }
                 title == "MORU님과 딱 맞는 루틴" -> feedRoutines.filter { it.authorName == "MORU" }
                 title == "이 루틴과 비슷한 루틴" -> {
-                    // 여기서는 어떤 루틴 기준인지 알 수 없으므로,
-                    // 예시로 '운동' 또는 '명상' 태그를 가진 루틴들을 보여줍니다.
                     feedRoutines.filter { it.tags.contains("운동") || it.tags.contains("명상") }
                 }
