+++ app/src/main/java/com/konkuk/moru/presentation/routinefeed/component/Routine/TitledRoutineSection.kt
@@ -1,4 +1,4 @@
-package com.konkuk.moru.presentation.routinefeed.component
+package com.konkuk.moru.presentation.routinefeed.component.Routine
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
